Refactor the books logic into a separate directory
